Final Commit
This is the final commit to this repo. After this, the repo is ready to go in production. Although with this commit attaching some .md files which consist of potential changes that can be made. Although even after those changes this repo can be easily and effectively sent into the production.